I removed (well commented out) the code to convert newlines and pars from BDSKConverter.  The stringByTexify and stringByDeTexify is not the right way to do this (which I use for annotated bibliographies) because they are called often and by methods that really break with the \newline in there!
This fixes RSS exporting again which was broken in the v0.9 release.  See Bug https://sourceforge.net/tracker/index.php?func=detail&aid=875617&group_id=61487&atid=497423

Of course it also breaks newline and par conversion.  I'll work on a proper way to do this for the next release but since we don't advertise the feature I think this is the lesser of two evils ;)