Fix some character conversion issues (945779, 945596, 945590).
Add one-way conversion of en-dash and em-dash, since those are
common when copy-pasting into annote.  They are converted to -- and ---,
respectively, in the Roman-->TeX dictionary, but since we scan for a {\
sequence in stringByDeTeXifyingString, the conversion is permanent.

It was necessary to add the characters to the character set as a string,
since they are out of the range we scan for.  This looks like a kludge, but
was the cleanest way I could see, without moving to an external file.