Add new files for autocompletion input manager preferences.
The input manager will now be bundled with Bibdesk in the Resources
directory, and the user can install it from Bibdesk's prefs.
The input manager needs to exist in the build products directory (along
with the Omni frameworks).  There are corresponding changes in the
input manager's project to bundle it properly for installation.

I didn't add a separate target for building the input manager from the Bibdesk.pbxproj.