Preliminary support for interacting with TextExtras1.7 and its complete: bug.  Requires the user to setup KeyBinding.dict with a binding for the BD_complete: selector, which should always work.
If the TextExtras complete: selector is found, the AppKit one is used instead.

I don't like my comparison of imp1 > imp2, although it seems to work; Mike Ferris suggested using mach-o API to get the AppKit base address and the addresses of the two implementations, but I can't figure out how to do this.