An attempt to solve the problem with the 'publication date' in AppleScript.
Basically it seems a bad idea to have that property because of all the conversion between an NSCalendarDate and the Month and Year fields needed behind the scenes.

Instead, we add publication month and publication year properties to give direct access to the relevant fields.

Please try this, there may still be some problems.