Fleshed out the drag and drop for collections.
Now you can drag items from the table to the source list and either
create a new collection or drop it into a current one.

Deleting pubs removes them from collections.

Nothing to do with collections is undoable at the moment.

Pressing delete on the pubs table when not viewing the library is
broken - it doesn't delete, but it should remove from the selected
collection and it doesn't. There's time for that later.

A few clean-up code refactorings and documentation additions, mainly in BDSKDragOutlineView and BibDocument.