Preparing to implement RFE 1574178, which calls for more flexibility in the cite command (eg \cite{} but also [#key][]).
The RFE calls for the InputManager (autocompletion) to handle something like [#citekey][] and hey, I'm using MultiMarkdown so I'd like this too.  First thought I should get it working in the main application before tackling the InputManager.

This check in just removes the unused BibItem:citation (that was confusing me) and standardizes on using BibDocument_DataSource:writePublications to generate both the dropped (or copied) string, and the preview while dragging. Before we manually generated the preview while dragging (adding in the "\") which was really confusing me.