Testing support for FileView.framework.  Currently BibDocument UI functionality is fairly complete, but the linked-file pref is overriden to use the FileView.
BibEditor support is bogus; it has its own ephemeral ivar to test drag-and-drop.

BibItem support is lame; it has an ivar whose name and function dates back
to the original datasource methods, which used a file path.  Needs to keep an
array of BDSKFile objects instead.

Splitview category in BibEditor should be moved to a subclass or use
method swizzling.