Add support for creating BDSKAsynchronousDOServer such that the server is fully set up by the time -init returns.
Add a new method for creating the server asynchronously, since we need that for a couple of the objects.

Modify subclasses to create ivars before initializing groups to avoid race conditions.  Document this.

Remove the hack of calling _cmd on a delay, which could lead to an endless loop.

Make server flags an opaque struct.