Changing Finder labels should be enabled for multiple selections.  Non-file URLs are ignored as the comments indicate.
Set count to 0 instead of an if to lose some indentation.  Remove a break and
some comments that appear to be obsolete.

Restore URLsFromPasteboard, which is to contain all URLs read from the pboard
and avoid duplication in the NSPasteboard call.  Removal of that broke webloc
addition and caused addition of file: and http: URLs when dropping from the 
Finder.  Possibly there are other ways to do this, but they should be tested
before committing.