Add segmented controls to status bar to choose preview styles. Also allow templates in the side pane and a fileview in the bottom pane.
Drop the custom details/notes/abstract previews, that should be done through templates.

Use a custom segmented control to fit in the status bar. 

Change preview style menu accordingly, and duplicate it for the side pane. 

Use per-document settings for the preview. For now not saved to xattrs or prefs, as we may still change the enum.