Add a concrete subclass of NSTask, since Apple's implementation is not thread safe.  Race conditions exist which can prevent waitUntilExit (and presumably other methods) from functioning properly; this has caused sporadic problems for years on multiprocessor systems.
Use the new class in the TeX task and script groups, both of which use NSTask on background threads.  This allows removal of the waitUntilExit runloop hack in the TeX task; child processes should no longer become detached, and terminationStatus will not raise an exception after waitUntilExit returns.

Thanks to Justin C. Walker for reporting this bug and for testing the fixes.